added comments

+++ src/components/interviews/MyInterviews.js
@@ -5,7 +5,7 @@ import { Button } from 'react-bootstrap'
 import InterviewCard from "./InterviewCard"
 class MyInterviews extends Component {
     state = {
-        interviews:[]
+        interviews: []
     }
 
     handleInputChange = (evt) => {
@@ -15,10 +15,10 @@ class MyInterviews extends Component {
     }
 
     componentDidMount() {
-        // gets all companies for the dropdown
+        // gets all interviews for the specific user
         APIManager.getAllAuth("interviews?applicant=true")
             .then((interviews) => {
-                // sets the companies in state
+                // sets the interviews in state
                 this.setState({
                     interviews: interviews
                 })
@@ -26,28 +26,33 @@ class MyInterviews extends Component {
     }
 
     deleteInterview = (id) => {
-        APIManager.delete("interviews", id)
-        .then(() => {
-            APIManager.getAllAuth("interviews?applicant=true")
-            .then((interviews) => {
-                // sets the companies in state
-                this.setState({
-                    interviews: interviews
-                })
-            })
+        // confirm the user wants to delete the interview
+        if (window.confirm("Are you sure you want to delete this interview?")) {
+            // make a DELETE request to the DB for the selected interview
+            APIManager.delete("interviews", id)
+                .then(() => {
+                    // get all interviews again
+                    APIManager.getAllAuth("interviews?applicant=true")
+                        .then((interviews) => {
+                            // update state with interviews
+                            this.setState({
+                                interviews: interviews
+                            })
+                        })
 
-        })
+                })
+        }
     }
 
-    
+
     render() {
 
         return (
             <>
-            <Link to="/interview/new"><Button>New Interview</Button></Link>
-            {this.state.interviews.map((interview) => {
-                return <InterviewCard {...this.props} key={interview.id} interview={interview} deleteInterview={this.deleteInterview}/>
-            })}
+                <Link to="/interview/new"><Button>New Interview</Button></Link>
+                {this.state.interviews.map((interview) => {
+                    return <InterviewCard {...this.props} key={interview.id} interview={interview} deleteInterview={this.deleteInterview} />
+                })}
 
             </>
         )
